Infer the server URL from the browser address
This simplifies deployment when Sealog client and server are hosted on the same
machine. It also handles the case where Sealog might be accessed from two
subnets, such as JasonNet and ScienceNet.

The change is a little obscure because we need to test to see if `window` is
defined. It is in the browser environment, but not in the Node environment the
application is built in. However, the client_config.js file is included by the
webpack config script, which executes in the Node environment, at which point
the attempt to access `window` fails.

Thanks to Muhammad Talha Akbar @dev-talha-akbar for pointing out the issue and
its solution.