Allow deployment-specific configuration of image URLs
The Sealog server serves images at URLs based on IMAGE_ROUTE. However, a design
shortcoming is that the API does not reference images relative to IMAGE_ROUTE
but to IMAGE_PATH, the directory where the files are actually stored.

This means that the client needs to know how to translate between the server's
internal implementation details (i.e., how it stores images) and its public
interface (i.e., how it serves them).

Ideally this could be fixed by simply changing the server to always reference
images based on the public URL.

But this is a risky change because services running on the server host might
expect the Sealog API to return a path on the local filesystem. These services
will have to be audited for such cases and modified to retrieve images over the
API.

In the meantime, this patch allows a configurable get_image_url() function that
takes a path relative to IMAGE_PATH and transforms it to a URL where the image
can be retrieved.

This enables an important use case, which is requesting images through a
caching image proxy to reduce ship-to-shore bandwidth use.