Use full path for image files
The basename is not enough to identify a given image unless all images from all
cruises and lowerings are stored in a single, flat directory. Our images are
sorted by cruise and by camera.

If the image_path begins with a /, @hapi/inert will treat it as an absolute path
outside of the image directory and return a 403 Forbidden error. We strip the
leading / off to avoid this case.