Refactor cruise editors to improve code re-use
Previously there were two separate forms for creating and updating a cruise,
duplicating most of the rendering and validation code. This change combines
them both into a single CruiseEditor component.

The application also tracked the currently active cruise in the Redux store,
which is mostly unnecessary: we can just pass it to components as a prop. This
change also removes the active cruise from the Redux store.

This also includes some minor refactoring such as adding helpers for evaluating
user permissions and reducing indentation levels.