Refactor event aux_data cards to improve code re-use
Previuosly, the virtually the same rendering code was duplicated between the
event history, event details modal, and lowering replay components.

This change introduces shared components, namely the DataCard component that
delegates to a specialized component for a given data_source, or falls back
on a default tabular card. This allows us to more easily customize how data
from certain sources are displayed across the entire app.

The DataCardGrid component outputs cards for each of an event's associated
aux_data objects. This will probably be replaced with an EventPreview class.

Currently this change introduces a regression in the tidiness of the UI.