Build a portable application that can be served from any path
Previously, the ROOT_PATH configuration option had to be hardcoded to the path
under which Sealog would be served, e.g., /sealog/. This added complexities to
how the build process worked, how Nginx had to be configured, etc.

  - ROOT_PATH is now derived automatically from the URL at which the JS bundle
    is loaded.

  - Sealog is now built at Docker container build time, not at run time, so
    the container now starts instantly.

  - The Nginx config no longer has to be patched to match ROOT_PATH.

There is a regression in that the configuration files are now baked in at
build time and cannot be overridden. This will be resolved.