Load the configuration scripts dynamically
Previously, the client_config.js and map_tilelayers.js configuration files
were baked into the application, which meant we had to re-build the whole
application each time the config changed.

This also fixes the Docker container so that the config is customizable again.

Because Create React App doesn't allow us to modify the Webpack config to set
the config scripts as externals, instead we introduce shim packages which
simply re-export the config settings from global variables.

The config files now live in config/ and not src/.